// src/app/models/activity-data.ts
export class ActivityData {
  number: number;
  altitude: number;
  altitudeDiffDownhill: number;
  altitudeDiffUphill: number;
  cadence: number;
  calories: number;
  distanceAbsolute: number;
  distance: number;
  distanceDownhill: number;
  distanceUphill: number;
  heartRate: number;
  intensityZone: number;
  rotations: number;
  speed: number;
  speedTime: number;
  targetZone: number;
  temperature: number;
  trainingTime: number;
  trainingTimeAbsolute: number;
  trainingTimeDownhill: number;
  trainingTimeUphill: number;
}